move routes to main blueprint
